AP1014_検査結果を登録するAPIを実装

+++ Backend/src/Wellship.WebAPI/ResultCollector/Usecases/HomeMenuUsecase.cs
@@ -31,7 +31,7 @@ public class HomeMenuUsecase : IHomeMenuUsecase
         int? status = placeScheduleId switch
         {
             null => null,
-            not null => (await _placeScheduleRepository.GetPlaceScheduleLockingStatusAsync((Guid)placeScheduleId)).Status
+            not null => (int)(await _placeScheduleRepository.GetPlaceScheduleLockingStatusAsync((Guid)placeScheduleId)).Status
         };
 
         // 機能ごとの利用可能条件の設定
